Add calculatePercentageDifference function to utils module
